+++ src/App/Documentation/utils/DocHeading.js
@@ -3,7 +3,7 @@ import React from "react";
 // TODO: overkill? [EH]
 const _generateTitle = path => path.match(/[^/]*$/)[0]
     .split("-")
-    .map(word => word.charAt(0).toUpperCase() + word.substr(1))
+    .map((word, i) => i > 0 ? word : word.charAt(0).toUpperCase() + word.substr(1))
     .join(" ");
 
 const DocHeading = ({ location }) => {