+++ src/App/components/FormComponents/index.js
@@ -1,0 +1,11 @@
+import Checkbox from "./Checkbox";
+import Radio from "./Radio";
+import Togglebox from "./Togglebox";
+
+export default {
+    Checkbox,
+    Radio,
+    Togglebox
+};
+
+export { Checkbox, Radio, Togglebox };